add split_scomm.c/split_scomm.h into helix/rev1
