+++ keyboards/helix/rev1/serial_config.h
@@ -8,9 +8,6 @@
 #define SERIAL_PIN_MASK _BV(PD0)
 #define SERIAL_PIN_INTERRUPT INT0_vect
 
-#define SERIAL_SLAVE_BUFFER_LENGTH MATRIX_ROWS/2
-#define SERIAL_MASTER_BUFFER_LENGTH 1
-
 /// #error rev1 serial config
 
 #endif /* SOFT_SERIAL_CONFIG_H */
